rename all the things, lint, serialize errors passed to catch all as JSON
