+++ .eslintrc.yml
@@ -1,0 +1,40 @@
+env:
+  browser: true
+  es6: true
+extends: 'eslint:recommended'
+parser: babel-eslint
+parserOptions:
+  ecmaFeatures:
+    experimentalObjectRestSpread: true
+    jsx: true
+  sourceType: module
+plugins:
+  - react
+globals:
+  process:
+    - true
+  exports:
+    - true
+  require:
+    - true
+  Buffer:
+    - true
+rules:
+  indent:
+    - error
+    - 2
+  linebreak-style:
+    - error
+    - unix
+  quotes:
+    - error
+    - single
+  semi:
+    - error
+    - never
+  react/jsx-uses-react:
+    - error
+  react/jsx-uses-vars:
+    - error
+  no-console:
+    - off
