--- lambda/SetupCustomDomain/2_DNSHostedZoneVerifier/package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "native-dns": "^0.7.0",
-    "parse-domain": "^2.0.0"
-  }
-}
