--- lambda/SetupCustomDomain/5_ACMCertificateCreator/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "aws-sdk": "^2.188.0"
-  }
-}
