--- lambda/SetupCustomDomain/7_CloudFrontDistributionCreator/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "node-uuid": "^1.4.8"
-  }
-}
