+++ package.json
@@ -1,0 +1,8 @@
+{
+  "devDependencies": {
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.6.1",
+    "eslint-plugin-react": "^7.11.1",
+    "prettier-eslint-cli": "^4.7.1"
+  }
+}
