+++ src/lambdas/dynamodb_stream_to_sns_trigger.js
