+++ src/lambdas/setup_custom_domain/09_cloudfront_distribution_verifier.js
@@ -1,0 +1,107 @@
+const AWS = require('aws-sdk')
+AWS.config.update({ region: process.env.AWS_REGION })
+const dynamodb = new AWS.DynamoDB()
+const cloudfront = new AWS.CloudFront()
+
+class SkipToEndWithSuccessError extends Error {}
+
+const getItem = domainName => {
+  return new Promise((resolve, reject) => {
+    const params = {
+      TableName: process.env.DYNAMODB_TABLE,
+      Key: {
+        DomainName: {
+          S: domainName
+        }
+      }
+    }
+    dynamodb.getItem(params, (err, data) => {
+      if (err) {
+        reject(err)
+      } else if (Object.keys(data).length === 0) {
+        reject(new Error('domain_name not found'))
+      } else {
+        resolve({ item: data.Item })
+      }
+    })
+  })
+}
+
+const verifyDistribution = ({ item }) => {
+  return new Promise((resolve, reject) => {
+    const params = {
+      Id: item.CloudFrontDistributionID.S
+    }
+
+    cloudfront.getDistribution(params, (err, data) => {
+      if (err) {
+        reject(err)
+      } else {
+        const status = data.Distribution.Status
+        if (status === 'Deployed') {
+          resolve({ item, distribution: data })
+        } else {
+          reject(new Error(`CloudFront distribution status is ${status}`))
+        }
+      }
+    })
+  })
+}
+
+const getUpdateItemParams = ({ item, distribution }) => {
+  return new Promise(resolve => {
+    const updateItemParams = {
+      TableName: process.env.DYNAMODB_TABLE,
+      Key: {
+        DomainName: {
+          S: item.DomainName.S
+        }
+      },
+      UpdateExpression:
+        'SET CloudFrontDistributionVerifiedAt=:CloudFrontDistributionVerifiedAt',
+      ExpressionAttributeValues: {
+        ':CloudFrontDistributionVerifiedAt': {
+          N: `${Date.now()}`
+        }
+      }
+    }
+    resolve({ item, updateItemParams })
+  })
+}
+
+const updateItem = ({ item, updateItemParams }) => {
+  return new Promise((resolve, reject) => {
+    dynamodb.updateItem(updateItemParams, (err, data) => {
+      if (err) {
+        reject(err)
+      } else {
+        resolve()
+      }
+    })
+  })
+}
+
+exports.handler = (event, _context, callback) => {
+  const { domainName } = event
+  const success = () => callback(null, { domainName })
+  const failure = err => {
+    const serializedError = JSON.stringify({
+      message: err.message,
+      domainName: DomainName
+    })
+    callback(new Error(serializedError))
+  }
+
+  getItem(domainName)
+    .then(verifyDistribution)
+    .then(getUpdateItemParams)
+    .then(updateItem)
+    .then(() => success())
+    .catch(error => {
+      if (error instanceof SkipToEndWithSuccessError) {
+        success()
+      } else {
+        failure(error)
+      }
+    })
+}
