+++ src/lambdas/setup_custom_domain/10_success_handler.js
@@ -1,0 +1,75 @@
+const AWS = require('aws-sdk')
+AWS.config.update({ region: process.env.AWS_REGION })
+const dynamodb = new AWS.DynamoDB()
+
+const getItem = domainName => {
+  return new Promise((resolve, reject) => {
+    const params = {
+      TableName: process.env.DYNAMODB_TABLE,
+      Key: {
+        DomainName: {
+          S: domainName
+        }
+      }
+    }
+    dynamodb.getItem(params, (err, data) => {
+      if (err) {
+        reject(err)
+      } else if (Object.keys(data).length === 0) {
+        reject(new Error('domain_name not found'))
+      } else {
+        resolve({ item: data.Item })
+      }
+    })
+  })
+}
+
+const getUpdateItemParams = ({ item }) => {
+  return new Promise(resolve => {
+    const updateItemParams = {
+      TableName: process.env.DYNAMODB_TABLE,
+      Key: {
+        DomainName: {
+          S: item.DomainName.S
+        }
+      },
+      UpdateExpression: 'SET SetupVerifiedAt=:SetupVerifiedAt',
+      ExpressionAttributeValues: {
+        ':SetupVerifiedAt': {
+          N: `${Date.now()}`
+        }
+      }
+    }
+    resolve({ item, updateItemParams })
+  })
+}
+
+const updateItem = ({ item, updateItemParams }) => {
+  return new Promise((resolve, reject) => {
+    dynamodb.updateItem(updateItemParams, (err, data) => {
+      if (err) {
+        reject(err)
+      } else {
+        resolve()
+      }
+    })
+  })
+}
+
+exports.handler = (event, _context, callback) => {
+  const { domainName } = event
+  const success = () => callback(null, { domainName })
+  const failure = err => {
+    const serializedError = JSON.stringify({
+      message: err.message,
+      domainName: DomainName
+    })
+    callback(new Error(serializedError))
+  }
+
+  getItem(domainName)
+    .then(getUpdateItemParams)
+    .then(updateItem)
+    .then(() => success())
+    .catch(error => failure(error))
+}
