+++ src/lambdas/teardown_custom_domain/06_success_handler.js
@@ -1,0 +1,39 @@
+const AWS = require('aws-sdk')
+AWS.config.update({ region: process.env.AWS_REGION })
+const dynamodb = new AWS.DynamoDB()
+
+const deleteItem = domainName => {
+  return new Promise((resolve, reject) => {
+    const params = {
+      TableName: process.env.DYNAMODB_TABLE,
+      Key: {
+        DomainName: {
+          S: domainName
+        }
+      }
+    }
+    dynamodb.deleteItem(params, err => {
+      if (err) {
+        reject(err)
+      } else {
+        resolve()
+      }
+    })
+  })
+}
+
+exports.handler = (event, _context, callback) => {
+  const { domainName } = event
+  const success = () => callback(null, { domainName })
+  const failure = err => {
+    const serializedError = JSON.stringify({
+      message: err.message,
+      domainName
+    })
+    callback(new Error(serializedError))
+  }
+
+  deleteItem(domainName)
+    .then(() => success())
+    .catch(error => failure(error))
+}
