+++ src/package.json
@@ -1,0 +1,14 @@
+{
+  "devDependencies": {
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.6.1",
+    "eslint-plugin-react": "^7.11.1",
+    "prettier-eslint-cli": "^4.7.1"
+  },
+  "dependencies": {
+    "aws-sdk": "^2.325.0",
+    "native-dns": "^0.7.0",
+    "node-uuid": "^1.4.8",
+    "parse-domain": "^2.1.2"
+  }
+}
